update entity sets and types between removals
If both --entitiestokeep and --entitiestoexlcude parameters are used and --entitiestoexclude specifies entities that have already been removed by the removal step that handles the --entitiestokeep parameter, a NullReferenceException is thrown. This can be avoided by updating the entity sets and types after the first removal step so the second removal step only considers entity sets and types for removal that still exist.